+++ ghg_lulc/artifact.py
@@ -3,17 +3,10 @@ from typing import Dict, Tuple
 
 import numpy as np
 import pandas as pd
-from climatoology.base.artifact import (
-    Chart2dData,
-    RasterInfo,
-    _Artifact,
-    create_chart_artifact,
-    create_geotiff_artifact,
-    create_table_artifact,
-    ContinuousLegendData,
-)
+from climatoology.base.artifact import Chart2dData, RasterInfo, Artifact, ContinuousLegendData, ArtifactMetadata, Legend
+from climatoology.base.artifact_creators import create_raster_artifact, create_table_artifact, create_chart_artifact
 from climatoology.base.computation import ComputationResources
-from climatoology.utility.LULC import LabelDescriptor
+from climatoology.utility.lulc import LabelDescriptor
 from pydantic_extra_types.color import Color
 
 from ghg_lulc.utils import PROJECT_DIR, GhgStockSource, RASTER_NO_DATA_VALUE, EMISSION_PER_PIXEL_FACTOR
@@ -24,7 +17,7 @@ def create_classification_artifacts(
     lulc_after: RasterInfo,
     labels: Dict[str, LabelDescriptor],
     resources: ComputationResources,
-) -> Tuple[_Artifact, _Artifact]:
+) -> Tuple[Artifact, Artifact]:
     # Hack due to https://gitlab.gistools.geog.uni-heidelberg.de/climate-action/web-app/-/issues/114
     unknown_color = Color('gray')
     colormap = lulc_before.colormap
@@ -32,42 +25,46 @@ def create_classification_artifacts(
     legend = {v.name: Color(v.color) for _, v in labels.items()}
     legend['unknown'] = unknown_color
 
-    lulc_before_artifact = create_geotiff_artifact(
+    lulc_before_metadata = ArtifactMetadata(
+        name='Classification for period start',
+        filename='lulc_classification_before',
+        summary='LULC classification at the start of the analysis period.',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/02_LULC_classifications.md').read_text(
+            encoding='utf-8'
+        ),
+        primary=False,
+    )
+    lulc_before_artifact = create_raster_artifact(
         raster_info=RasterInfo(
             data=lulc_before.data.filled(fill_value=RASTER_NO_DATA_VALUE),
             crs=lulc_before.crs,
             transformation=lulc_before.transformation,
             colormap=colormap,
-            nodata=RASTER_NO_DATA_VALUE,
         ),
-        layer_name='Classification for period start',
-        caption='LULC classification at the start of the analysis period.',
+        legend=Legend(legend_data=legend),
+        resources=resources,
+        metadata=lulc_before_metadata,
+    )
+
+    lulc_after_metadata = ArtifactMetadata(
+        name='Classification for period end',
+        filename='lulc_classification_after',
+        summary='LULC classification at the end of the analysis period.',
         description=(PROJECT_DIR / 'resources/artifact_descriptions/02_LULC_classifications.md').read_text(
             encoding='utf-8'
         ),
-        legend_data=legend,
-        resources=resources,
-        filename='lulc_classification_before',
         primary=False,
     )
-
-    lulc_after_artifact = create_geotiff_artifact(
+    lulc_after_artifact = create_raster_artifact(
         raster_info=RasterInfo(
             data=lulc_after.data.filled(fill_value=RASTER_NO_DATA_VALUE),
             crs=lulc_after.crs,
             transformation=lulc_after.transformation,
             colormap=colormap,
-            nodata=RASTER_NO_DATA_VALUE,
         ),
-        layer_name='Classification for period end',
-        caption='LULC classification at the end of the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/02_LULC_classifications.md').read_text(
-            encoding='utf-8'
-        ),
-        legend_data=legend,
+        legend=Legend(legend_data=legend),
         resources=resources,
-        filename='lulc_classification_after',
-        primary=False,
+        metadata=lulc_after_metadata,
     )
 
     return lulc_before_artifact, lulc_after_artifact
@@ -79,14 +76,13 @@ def create_change_artifacts(
     ghg_stock: pd.DataFrame,
     emission_factors: pd.DataFrame,
     resources: ComputationResources,
-) -> Tuple[_Artifact, _Artifact]:
+) -> Tuple[Artifact, Artifact]:
     filled_change_data = change.data.filled(fill_value=RASTER_NO_DATA_VALUE)
     filled_change = RasterInfo(
         data=filled_change_data,
         crs=change.crs,
         transformation=change.transformation,
         colormap=change.colormap,
-        nodata=RASTER_NO_DATA_VALUE,
     )
     emission_factors['change'] = emission_factors.apply(
         lambda row: f'{row.utility_class_name_before} to {row.utility_class_name_after}', axis=1
@@ -94,15 +90,18 @@ def create_change_artifacts(
     lookup = emission_factors.set_index('change_id')['change'].to_dict()
     lookup[0] = 'No Change'
 
-    change_artifact = create_geotiff_artifact(
-        raster_info=filled_change,
-        layer_name='LULC Change',
-        caption='LULC changes within the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/03_LULC_change.md').read_text(encoding='utf-8'),
-        resources=resources,
+    change_metadata = ArtifactMetadata(
+        name='LULC Change',
         filename='LULC_change',
-        legend_data={lookup[change_id]: Color(color) for change_id, color in change.colormap.items()},
+        summary='LULC changes within the analysis period.',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/03_LULC_change.md').read_text(encoding='utf-8'),
         primary=False,
+    )
+    change_artifact = create_raster_artifact(
+        raster_info=filled_change,
+        resources=resources,
+        legend=Legend(legend_data={lookup[change_id]: Color(color) for change_id, color in change.colormap.items()}),
+        metadata=change_metadata,
     )
 
     filled_change_emissions_data = change_emissions.data.filled(fill_value=RASTER_NO_DATA_VALUE)
@@ -135,25 +134,30 @@ def create_change_artifacts(
         colormap=patched_emissions_colormap,
     )
 
-    patched_localised_emission_artifact = create_geotiff_artifact(
-        raster_info=patched_change_emissions,
-        layer_name='Localised carbon flows (tonnes per 100m²)',
-        caption='Carbon flows per pixel due to LULC change. Note: Transparent pixels within the area of interest '
+    patched_localised_emission_metadata = ArtifactMetadata(
+        name='Localised carbon flows (tonnes per 100m²)',
+        filename='LULC_change_emissions_patched',
+        summary='Carbon flows per pixel due to LULC change. Note: Transparent pixels within the area of interest '
         'represent areas where carbon flows from LULC change could not be estimated, either because the pixels could '
         'not be classified with sufficient confidence or they were classified as "permanent crops" or "water", for '
         'which no carbon stock values are available.',
         description=change_emission_description,
-        resources=resources,
-        filename='LULC_change_emissions_patched',
-        legend_data=ContinuousLegendData(
-            cmap_name='coolwarm',
-            ticks={
-                str(min(emission_factors.emission_factor) * EMISSION_PER_PIXEL_FACTOR): 0,
-                str(0): 0.5,
-                str(max(emission_factors.emission_factor) * EMISSION_PER_PIXEL_FACTOR): 1,
-            },
-        ),
         primary=True,
+    )
+    patched_localised_emission_artifact = create_raster_artifact(
+        raster_info=patched_change_emissions,
+        resources=resources,
+        legend=Legend(
+            legend_data=ContinuousLegendData(
+                cmap_name='coolwarm',
+                ticks={
+                    str(min(emission_factors.emission_factor) * EMISSION_PER_PIXEL_FACTOR): 0,
+                    str(0): 0.5,
+                    str(max(emission_factors.emission_factor) * EMISSION_PER_PIXEL_FACTOR): 1,
+                },
+            ),
+        ),
+        metadata=patched_localised_emission_metadata,
     )
 
     return change_artifact, patched_localised_emission_artifact
@@ -170,85 +174,103 @@ def patch_change_data(change_data: np.ndarray, orig_colormap: Dict[Number, Tuple
     return patched_data, patched_colormap
 
 
-def create_summary_artifact(summary_df: pd.DataFrame, resources: ComputationResources) -> _Artifact:
+def create_summary_artifact(summary_df: pd.DataFrame, resources: ComputationResources) -> Artifact:
+    summary_metadata = ArtifactMetadata(
+        name='Summary of results',
+        filename='summary',
+        summary='Gross emissions, gross sinks, and net emissions/sinks in the analysis period.',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/10a_summary.md').read_text(encoding='utf-8'),
+        primary=True,
+    )
     summary_artifact = create_table_artifact(
         data=summary_df,
-        title='Summary of results',
-        caption='Gross emissions, gross sinks, and net emissions/sinks in the analysis period.',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/10a_summary.md').read_text(encoding='utf-8'),
         resources=resources,
-        filename='summary',
-        primary=True,
+        metadata=summary_metadata,
     )
     return summary_artifact
 
 
-def create_area_info_artifact(area_info_df: pd.DataFrame, resources: ComputationResources) -> _Artifact:
-    summary_artifact = create_table_artifact(
-        data=area_info_df,
-        title='Summary of detected changes',
-        caption='Size of the area classified as a carbon source, carbon sink, and total LULC change during the period '
+def create_area_info_artifact(area_info_df: pd.DataFrame, resources: ComputationResources) -> Artifact:
+    area_info_metadata = ArtifactMetadata(
+        name='Summary of detected changes',
+        filename='area_info',
+        summary='Size of the area classified as a carbon source, carbon sink, and total LULC change during the period '
         'of analysis.',
         description=(PROJECT_DIR / 'resources/artifact_descriptions/10b_area_info.md').read_text(encoding='utf-8'),
-        resources=resources,
-        filename='area_info',
         primary=False,
+    )
+    summary_artifact = create_table_artifact(
+        data=area_info_df,
+        resources=resources,
+        metadata=area_info_metadata,
     )
     return summary_artifact
 
 
 def create_stock_artifact(
     stock_df: pd.DataFrame, stock_source: GhgStockSource, resources: ComputationResources
-) -> _Artifact:
+) -> Artifact:
+    stock_metadata = ArtifactMetadata(
+        name='Carbon stock values per class',
+        filename='stock',
+        summary=f'Carbon stock values for each class according to: {stock_source.value}',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/08_ghg_stocks.md').read_text(encoding='utf-8'),
+        primary=False,
+    )
     stock_artifact = create_table_artifact(
         data=stock_df,
-        title='Carbon stock values per class',
-        caption=f'Carbon stock values for each class according to: {stock_source.value}',
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/08_ghg_stocks.md').read_text(encoding='utf-8'),
         resources=resources,
-        filename='stock',
-        primary=False,
+        metadata=stock_metadata,
     )
     return stock_artifact
 
 
-def create_change_type_artifact(change_type_table: pd.DataFrame, resources: ComputationResources) -> _Artifact:
-    change_type_table_artifact = create_table_artifact(
-        data=change_type_table,
-        title='Change areas and carbon flows by LULC change type',
-        caption='Total change area by LULC change type (ha) and total carbon flows by '
+def create_change_type_artifact(change_type_table: pd.DataFrame, resources: ComputationResources) -> Artifact:
+    change_type_metadata = ArtifactMetadata(
+        name='Change areas and carbon flows by LULC change type',
+        filename='stats_change_type',
+        summary='Total change area by LULC change type (ha) and total carbon flows by '
         'LULC change type (tonnes) in the analysis period.',
         description=(PROJECT_DIR / 'resources/artifact_descriptions/09_stats_change_type.md').read_text(
             encoding='utf-8'
         ),
-        resources=resources,
-        filename='stats_change_type',
         primary=True,
+    )
+    change_type_table_artifact = create_table_artifact(
+        data=change_type_table,
+        resources=resources,
+        metadata=change_type_metadata,
     )
     return change_type_table_artifact
 
 
-def create_area_plot_artifact(area_data: Chart2dData, resources: ComputationResources) -> _Artifact:
+def create_area_plot_artifact(area_data: Chart2dData, resources: ComputationResources) -> Artifact:
+    area_data_metadata = ArtifactMetadata(
+        name='Change areas by LULC change type (ha)',
+        filename='area_plot',
+        summary='Change areas by LULC change type (ha) in the analysis period.',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/07_area_plot.md').read_text(encoding='utf-8'),
+        primary=True,
+    )
     area_data_artifact = create_chart_artifact(
         data=area_data,
-        title='Change areas by LULC change type (ha)',
-        caption='Change areas by LULC change type (ha) in the analysis period.',
         resources=resources,
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/07_area_plot.md').read_text(encoding='utf-8'),
-        filename='area_plot',
-        primary=True,
+        metadata=area_data_metadata,
     )
     return area_data_artifact
 
 
-def create_emission_plot_artifact(emission_data: Chart2dData, resources: ComputationResources) -> _Artifact:
+def create_emission_plot_artifact(emission_data: Chart2dData, resources: ComputationResources) -> Artifact:
+    emission_data_metadata = ArtifactMetadata(
+        name='Carbon flows by LULC change type (tonnes)',
+        filename='emission_plot',
+        summary='Carbon flows by LULC change type (tonnes) in the analysis period.',
+        description=(PROJECT_DIR / 'resources/artifact_descriptions/06_emission_plot.md').read_text(encoding='utf-8'),
+        primary=True,
+    )
     emission_data_artifact = create_chart_artifact(
         data=emission_data,
-        title='Carbon flows by LULC change type (tonnes)',
-        caption='Carbon flows by LULC change type (tonnes) in the analysis period.',
         resources=resources,
-        description=(PROJECT_DIR / 'resources/artifact_descriptions/06_emission_plot.md').read_text(encoding='utf-8'),
-        filename='emission_plot',
-        primary=True,
+        metadata=emission_data_metadata,
     )
     return emission_data_artifact