+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project mostly adheres to [Semantic Versioning](https://semver.org/spec
 
 ### Changed
 - Add teaser to plugin and restore the previous purpose
+- Update the plugin to use the climatoology version 7.0.0
 
 ## [2.3.3](https://gitlab.heigit.org/climate-action/plugins/ghg-emission-from-lulc-change/-/releases/2.3.3) - 2025-06-13